fix: send pgns in plain format
