+++ lib/canbus.js
@@ -24,7 +24,7 @@ const Parser = require('./fromPgn').Parser
 const _ = require('lodash')
 const CanDevice = require('./candevice')
 const spawn = require('child_process').spawn
-const { getPGNFromCanId, getCanIdFromPGN, actisenseSerialToBuffer } = require('./utilities')
+const { getPlainPGNs, getPGNFromCanId, getCanIdFromPGN, actisenseSerialToBuffer } = require('./utilities')
 
 const MSG_BUF_SIZE  			= 2000
 const CANDUMP_DATA_INC_3		= 3
@@ -227,36 +227,6 @@ CanbusStream.prototype.sendPGN = function (msg) {
       }
     }
   }
-}
-
-function getPlainPGNs(buffer) {
-  var res = []
-  var bucket = 0x40
-
-  var first = new Buffer(8)
-  first.writeUInt8(bucket++, 0)
-  first.writeUInt8(buffer.length, 1)
-  buffer.copy(first, 2, 0, 6)
-  res.push(first)
-
-  for ( var index = 6; index < buffer.length; index += 7 ) {
-    var next = new Buffer(8)
-    next.writeUInt8(bucket++, 0)
-    var end = index+7
-    var fill = 0
-    if ( end > buffer.length ) {
-      fill = end - buffer.length
-      end = buffer.length
-    }
-    buffer.copy(next, 1, index, end)
-    if ( fill > 0 ) {
-      for ( var i = end-index; i < 8; i++ ) {
-        next.writeUInt8(0xff, i)
-      }
-    }
-    res.push(next)
-  }
-  return res
 }
 
 function readLine(that, line) {
