+++ lib/toPgn.js
@@ -19,7 +19,7 @@ const _ = require('lodash')
 const BitStream = require('bit-buffer').BitStream
 const Int64LE = require('int64-buffer').Int64LE
 const Uint64LE = require('int64-buffer').Uint64LE
-const { getManufacturerCode, getIndustryCode, getCanIdFromYdwgPGN } = require('./utilities')
+const { getPlainPGNs, getManufacturerCode, getIndustryCode, getCanIdFromYdwgPGN } = require('./utilities')
 
 const RES_STRINGLAU = 'ASCII or UNICODE string starting with length and control byte'
 
@@ -296,8 +296,16 @@ function toYdwgRawFormat(pgn, data, dst=255) {
   //canId data
   //19F51323 01 02<CR><LF>
   pgn.dst = dst
-  canId = getCanIdFromYdwgPGN(pgn)
-  return canId.toString(16) + " " + parseInt(data, 2).toString(16)
+  const canId = getCanIdFromYdwgPGN(pgn)
+  return getPlainPGNs(data).map(buffer => {
+    return canId.toString(16) + " " + new Uint32Array(buffer)
+      .reduce(function(acc, i) {
+        acc.push(i.toString(16));
+        return acc;
+      }, [])
+      .map(x => (x.length === 1 ? "0" + x : x))
+      .join(" ")
+  })
 }
 
 function pgnToYdwgRawFormat(pgn) {
