+++ lib/utilities.js
@@ -274,6 +274,37 @@ function getDeviceClassName(code) {
   return deviceClassCodes[code]
 }
 
+function getPlainPGNs(buffer) {
+  var res = []
+  var bucket = 0x40
+
+  var first = new Buffer(8)
+  first.writeUInt8(bucket++, 0)
+  first.writeUInt8(buffer.length, 1)
+  buffer.copy(first, 2, 0, 6)
+  res.push(first)
+
+  for ( var index = 6; index < buffer.length; index += 7 ) {
+    var next = new Buffer(8)
+    next.writeUInt8(bucket++, 0)
+    var end = index+7
+    var fill = 0
+    if ( end > buffer.length ) {
+      fill = end - buffer.length
+      end = buffer.length
+    }
+    buffer.copy(next, 1, index, end)
+    if ( fill > 0 ) {
+      for ( var i = end-index; i < 8; i++ ) {
+        next.writeUInt8(0xff, i)
+      }
+    }
+    res.push(next)
+  }
+  return res
+}
+
+
 const defaultTransmitPGNs = [
   60928,
   59904,
@@ -304,3 +335,4 @@ module.exports.getManufacturerCode = getManufacturerCode
 module.exports.getDeviceClassName = getDeviceClassName
 module.exports.getDeviceClassCode = getDeviceClassCode
 module.exports.defaultTransmitPGNs = defaultTransmitPGNs
+module.exports.getPlainPGNs = getPlainPGNs
