+++ lib/ydwg02.js
@@ -78,8 +78,9 @@ Ydwg02Stream.prototype.sendPGN = function (pgn) {
   let now = Date.now()
   let lastSent = pgnsSent[pgn.pgn]
   if ( !lastSent || now - lastSent > rateLimit ) {
-    let msg = pgnToYdwgRawFormat(pgn)
-    this.sendString(msg)
+    pgnToYdwgRawFormat(pgn).forEach(raw => {
+      this.sendString(raw)
+    })
     pgnsSent[pgn.pgn] = now
   }
 }
@@ -98,8 +99,9 @@ Ydwg02Stream.prototype.sendYdwgPGN = function (msg) {
       let now = Date.now()
       let lastSent = pgnsSent[pgn.pgn]
       if ( !lastSent || now - lastSent > rateLimit ) {
-        let msg = pgnToYdwgRawFormat(pgn)
-        that.sendString(msg)
+        pgnToYdwgRawFormat(pgn).forEach(raw => {
+          this.sendString(raw)
+        })
         pgnsSent[pgn.pgn] = now
       }
     })
